requirements.txt
# django
Django
python-dotenv
django-storages
djangorestframework
django-filter
